Add solver

+++ src/Icfpc2023/Domain/Taste.cs
@@ -4,5 +4,5 @@ public class Taste
 {
     public Instrument Instrument { get; init; }
 
-    public float Value { get; init; }
+    public double Value { get; init; }
 }
